Extract healthcheck app DB tests into separate file
Makes it easier to maintain with these separated out.
